requirements.txt
python-dateutil
tqdm
sklearn
cython
humanize
pandas
fire
h5py
tabulate
openpyxl
psutil
